requirements.txt
numpy>=1.20.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
opencv-python>=4.5.0
pycolmap>=0.3.0
pyproj>=3.0.0
utm>=0.7.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
faiss-gpu>=1.7.0
exifread>=3.0.0
geopy>=2.2.0
r2d2-pytorch>=0.1.0